+++ Src/GloabalValues.h
@@ -87,8 +87,8 @@ struct Vertex {
 };
 
 struct ObjectDescription {
-
-	int texture_id;
+	
+	uint64_t texture_id;
 	uint64_t vertex_address;
 	uint64_t index_address;
